added unix timestamping to create user call

+++ backend/internal/models/auth_models.go
@@ -10,7 +10,7 @@ type SignupRequest struct {
 	Name      string `json:"name"`
 	Email     string `json:"email"`
 	Password  string `json:"password"`
-	CreatedAt int    `json:"created_at"`
+	CreatedAt int64  `json:"created_at"`
 }
 
 type User struct {
